% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{.abbrev}
\alias{.abbrev}
\title{Abbreviate names}
\usage{
.abbrev(x)
}
\description{
Manually specify field names for writing to shapefile instead of relying on
\code{sf:::abbreviate_shapefile_names()} (which uses \code{abbreviate()}) during save.
Names must contain fewer than 10 characters.
}
\author{
Alex Chubaty
}
\keyword{internal}
