% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.emptySP}
\alias{.emptySP}
\alias{.emptySPDF}
\title{Create emtpy/null SpatialPolygons* objects}
\usage{
.emptySP()

.emptySPDF()
}
\description{
Internal use only.
}
\section{Warning}{

Do not attempt to \code{print} or \code{show} these objects, as you'll get caught
in an inifinite loop that requires termitation of the R process!
}

\keyword{internal}
