% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,mpg-method}
\alias{export,grain-method}
\title{Export spatial data from MPG and GOC models}
\usage{
export(x, dirname = NULL, path = ".", rasterFormat = "GTiff",
  overwrite = FALSE, R = FALSE, vorBound = FALSE, ...)

\S4method{export}{mpg}(x, dirname = NULL, path = ".",
  rasterFormat = "GTiff", overwrite = FALSE, R = FALSE,
  vorBound = FALSE, ...)

\S4method{export}{grain}(x, dirname = NULL, path = ".",
  rasterFormat = "GTiff", overwrite = FALSE, R = FALSE,
  vorBound = FALSE, ...)
}
\arguments{
\item{x}{A \code{mpg} or \code{grain} object}

\item{dirname}{The name of a new directory to create.  If \code{NULL}
a directory with a name containing the date and time will be created.}

\item{path}{A path to where this new directory \code{dirname} should be created.
Defaults to the working directory.}

\item{rasterFormat}{The format for exported rasters.  See \code{\link{writeFormats}} for
options.  Defaults to GeoTiff (\code{rasterFormat='GTiff'}).  Use
\code{rasterFormat='raster'} to save \code{.grd} files in
native \code{raster} package format.}

\item{overwrite}{If directory already exists will overwrite existing files inside.
Defaults to \code{FALSE}.}

\item{R}{If \code{TRUE}, return the spatial objects that would be written to files.  Do not
write these files and ignore \code{dirname, path, rasterFormat, overwrite}
parameters.  This is useful for visualization using R plotting functions, or
spatial analysis within R. Defaults to \code{FALSE}}

\item{vorBound}{Specify whether to create a raster with the boundaries of the Voronoi
polygons \code{=1} and the remainder \code{=NA}.  This may be useful
for visualizing relationships among polygons in a grain of connectivity.
This can add time to the export on very large rasters.
Defaults to \code{FALSE}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the path to the folder created.

  Side effect of exporting files representing raster and vector spatial data
  in the object.

  Please note that for vector data export the attribute name is limited
  to 8 characters in shape files.  See the tables below for the abbreviations
  used and their meaning.

  \strong{Exported from \code{mpg} objects:}

  \code{nodes, linksCentroid, linksPerim} are shape files giving the
  locations of the patch centroids, links among centroids, and links
  among perimeters of patches respectively.  \code{patchId, voronoi} are
  rasters giving the patch identifier of the patch, or of the patch that
  the Voronoi polygon refers to.  \code{lcpPerimWeight, lcpLinkId} give
  the weight in cost surface units of the shortest paths between perimeters,
  and the identifiers of those links respectively.  \code{vorBound} gives
  the boundaries of the Voronoi polygons (if specified).

  Description of node (vertex) and link (edge) weights in \code{mpg}
  objects  and their corresponding attribute names in the
  shape files created.

\tabular{llll}{
  \strong{type} \tab \strong{MPG name} \tab \strong{SHP name} \tab \strong{Description}\cr
  node \tab patchId \tab patchId \tab Patch id from patchId raster \cr
  node \tab patchArea \tab patchA \tab Area of patch \cr
  node \tab patchEdgeArea \tab patchEA \tab Edge area of patch \cr
  node \tab coreArea \tab coreA \tab Area excluding edge of patch \cr
  node \tab centroidX \tab ctrX \tab Centroid of the patch (X)\cr
  node \tab centroidY \tab ctrY \tab Centroid of the patch (Y)\cr
  link \tab e1 \tab e1 \tab Id of first patch at end of link\cr
  link \tab e2 \tab e2 \tab Id of second patch at end of link\cr
  link \tab linkId \tab linkId \tab Link id from lcpLinkId raster\cr
  link \tab lcPerimWeight \tab lcpPerWt \tab Cost length of link from patch perimeters\cr
  link \tab startPerimX \tab strtPerX \tab Coordinate of link endpoint on first patch (X)\cr
  link \tab startPerimY \tab strtPerY \tab Coordinate of link endpoint on first patch (Y)\cr
  link \tab endPerimX \tab endPerX \tab Coordinate of link endpoint on second patch (X)\cr
  link \tab endPerimY \tab endPerY \tab Coordinate of link endpoint on second patch (Y)\cr
}

  \strong{Exported from \code{grain} objects:}

  \code{nodes, linksCentroid} are shape files giving the locations of the
  Voronoi polygon centroids and links among them respectively. \code{voronoi}
  are rasters gives the polygon identifier of each cluster of patches.
  \code{vorBound} gives the boundaries of the Voronoi polygons (if specified).

  Description of node (vertex) and link (edge) weights in \code{grain}
  objects  and their corresponding attribute names in the
  shape files created.

\tabular{llll}{
  \strong{type} \tab \strong{GOC name} \tab \strong{SHP name} \tab \strong{Description}\cr
  node \tab polygonId \tab polyId \tab Polygon id from grain voronoi raster \cr
  node \tab polygonArea \tab polyA \tab Area of polygon from grain voronoi raster \cr
  node \tab totalPatchArea \tab patchA \tab Total area of all patches in polygon\cr
  node \tab totalPatchEdgeArea \tab patchEA \tab Total area of all patch edges in polygon\cr
  node \tab totalCoreArea \tab coreA \tab Total area of patches in poylgon excluding edges\cr
  node \tab centroidX \tab ctrX \tab Centroid of the polygon (X)\cr
  node \tab centroidY \tab ctrY \tab Centroid of the polygon (Y)\cr
  link \tab e1 \tab e1 \tab Id of first patch at end of link\cr
  link \tab e2 \tab e2 \tab Id of second patch at end of link\cr
  link \tab maxWeight \tab maxWt \tab The maximum weight of all links connecting patches between polygons\cr
  link \tab linkIdMaxWeight \tab maxWt \tab The link id of that maximum weight link (lcpLinkId)\cr
  link \tab minWeight \tab min \tab The minimum weight of all links connecting patches between polygons\cr
  link \tab linkIdMinWeight \tab minWt \tab The link id of that minimum weight link (lcpLinkId)\cr
  link \tab medianWeight \tab medWt \tab The median weight of all links connecting patches between polygons\cr
  link \tab meanWeight \tab meanWT \tab The minimum weight of all links connecting patches between polygons\cr
  link \tab numlinksWeight \tab numEWt \tab The number of links connecting patches between polygons\cr
  link \tab eucCentroidWeight \tab eucCtrWt \tab The Euclidean distance between centroids of polygons\cr
}
}
\description{
This function automates the export of raster and vector spatial data from
\code{mpg} and \code{grain} objects.  By default it places them in a new directory,
unless an existing one is specified with \code{overwrite=TRUE}.

It can also be used to process \code{mpg} and \code{grain} objects to
produce R spatial objects that are convenient for plotting or analysis within R.
Use \code{R=TRUE} in which case all parameters related to file export
are ignored.  (Default \code{R=FALSE})

The \code{raster} package \code{\link{writeRaster}} is used for rasters,
and \code{\link{writeOGR}} in the \code{rgdal} package is used to
export ESRI compatible shape files.
}
\examples{
\dontrun{
library(raster)

## Load raster landscape
tiny <- raster(system.file("extdata/tiny.asc", package = "grainscape"))

## Create a resistance surface from a raster using an is-becomes reclassification
tinyCost <- reclassify(tiny, rcl = cbind(c(1, 2, 3, 4), c(1, 5, 10, 12)))

## Produce a patch-based MPG where patches are resistance features=1
tinyPatchMPG <- MPG(cost = tinyCost, patch = tinyCost == 1)

## Extract a representative subset of 5 grains of connectivity
tinyPatchGOC <- GOC(tinyPatchMPG, nThresh = 5)

## Export rasters and vectors of the MPG to the current working directory
export(tinyPatchMPG)

## Export raster and vectors of a grain to a specified directory
export(grain(tinyPatchGOC, 2), dirname = 'tiny_goc_thresh2')
}

}
\seealso{
\code{\link{MPG}}, \code{\link{GOC}}, \code{\link{grain}}
}
\author{
Paul Galpern and Alex Chubaty
}
