% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{$}
\alias{$}
\alias{$,goc-method}
\alias{$<-}
\alias{$<-,goc-method}
\alias{$,mpg-method}
\alias{$<-,mpg-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{$}{goc}(x, name)

\S4method{$}{goc}(x, name) <- value

\S4method{$}{mpg}(x, name)

\S4method{$}{mpg}(x, name) <- value
}
\arguments{
\item{x}{A \code{simList} object from which to extract element(s) or
in which to replace element(s).}

\item{name}{A literal character string or a \code{\link{name}}.}

\item{value}{Any R object.}
}
\description{
Extract or Replace Parts of an Object
}
