% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{goc-class}
\alias{goc-class}
\title{The \code{goc} class}
\description{
The \code{goc} class
}
\section{Slots}{

\describe{
\item{\code{voronoi}}{A \code{RasterLayer} describing the regions of proximity in
resistance units around the focal patches or points.}

\item{\code{summary}}{A summary of the the grains of connectivity generated and
their properties.}

\item{\code{th}}{A list giving the GOC graph at each threshold.

Each element of \code{th} contains a \code{goc} object giving the GOC graph
as class \code{\link{igraph}}.
Vertex attributes describes qualities of each polygon including the coordinates
of each polygon centroid, the area of these polygons, and the original patch
IDs in the MPG that are included in each polygon.
All areal measurements are given as raster cell counts.
A variety of edge attributes are also given in the GOC graph.
See \code{\link{distance}} for more information.}
}}

\author{
Alex Chubaty and Paul Galpern
}
