% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOC.R
\docType{methods}
\name{GOC}
\alias{GOC}
\alias{GOC,mpg-method}
\title{Produce a grains of connectivity model at multiple scales (patch-based or lattice GOC)}
\usage{
GOC(x, ...)

\S4method{GOC}{mpg}(x, nThresh = NULL, doThresh = NULL,
  weight = "lcpPerimWeight", sp = FALSE, verbose = 0, ...)
}
\arguments{
\item{x}{A \code{mpg} object produced by \code{\link{MPG}}.
For lattice GOC \code{MPG} must be run with patch set as an integer value.}

\item{...}{Additional arguments (not used).}

\item{nThresh}{Optional.  An integer giving the number of thresholds (or scales)
at which to create GOC models.  Thresholds are selected to produce
a maximum number of unique grains (i.e., models).
\code{nThresh} thresholds are also approximately evenly spread
between 0 and the threshold at which all patches or focal points
on the landscape are connected.  This is a simple way to get
a representative subset of all possible GOC models.
Provide either \code{nThresh} or \code{doThresh} not both.}

\item{doThresh}{Optional.  A vector giving the link thresholds at which to create GOC models.
Use \code{link{threshold}} to identify thresholds of interest.
Provide either \code{nThresh} or \code{doThresh} not both.}

\item{weight}{A string giving the link weight or attribute to use for threshold.
\code{"lcpPerimWeight"} uses the accumulated resistance or least-cost
path distance from the perimeters of patches as the link weight.
\code{"eucPerimWeight"} use the Euclidean distance from the
perimeters of patches as the link weight.}

\item{sp}{Logical.  If \code{TRUE} the \code{rgeos} package is used to create
a vector of class \code{\link{SpatialPolygonsDataFrame}} describing
the finest grain of connectivity. This is very useful for visualizing
grains of connectivity models, especially for print purposes.
Equally, using the \code{maptools} or \code{rgdal} packages these
polygons can be exported as shapefiles for use in other GIS applications.
But, please see details.}

\item{verbose}{Set \code{verbose=0} for no progress information to console.}
}
\value{
A \code{\link[=goc-class]{goc}} object.
}
\description{
Produce a grains of connectivity (GOC) model at multiple scales (resistance thresholds)
by scalar analysis.
Patch-based or lattice GOC modelling can be done with this function.
}
\details{
This function can take a long time to run when \code{sp = TRUE}.
Time taken is dependent on the dimensions of the \code{mpg$voronoi} raster.
}
\note{
Researchers should consider whether the use of a patch-based GOC or a lattice
GOC model is appropriate based on the patch-dependency of the organism under study.
Patch-based models make most sense when animals are restricted to, or dependent on, a resource patch.
Lattice models can be used as a generalized and functional approach to scaling resistance surfaces.

See \code{\link{MPG}} for warning related to areal measurements.
}
\examples{
\dontrun{
library(raster)

## Load raster landscape
tiny <- raster(system.file("extdata/tiny.asc", package = "grainscape"))

## Create a resistance surface from a raster using an is-becomes reclassication
tinyCost <- reclassify(tiny, rcl = cbind(c(1, 2, 3, 4), c(1, 5, 10, 12)))

## Produce a patch-based MPG where patches are resistance features=1
tinyPatchMPG <- MPG(cost = tinyCost, patch = (tinyCost == 1))

## Extract a representative subset of 5 grains of connectivity
tinyPatchGOC <- GOC(tinyPatchMPG, nThresh = 5)

## Examine the properties of the GOC graph of grain 3 of 5
print(tinyPatchGOC@th[[3]]$goc, vertex = TRUE, edge = TRUE)

## Extract specified grains of connectivity and produce a vector SpatialPolygons
## representation of the finest grain of connectivity (Threshold = 0)
tinyPatchGOC <- GOC(tinyPatchMPG, doThresh = c(0, 20, 40), sp = TRUE)
}

}
\references{
Fall, A., M.-J. Fortin, M. Manseau, D. O'Brien. (2007) Spatial graphs: Principles and applications for habitat connectivity. Ecosystems 10:448:461.

Galpern, P., M. Manseau, P.J. Wilson. (2012) Grains of connectivity: analysis at multiple spatial scales in landscape genetics. Molecular Ecology 21:3996-4009.
}
\seealso{
\code{\link{MPG}}, \code{\link{grain}},
         \code{\link{distance}}, \code{\link{point}}
}
\author{
Paul Galpern
}
