% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{grain-class}
\alias{grain-class}
\title{The \code{grain} class}
\description{
The \code{grain} class
}
\section{Slots}{

\describe{
\item{\code{voronoi}}{A \code{RasterLayer} describing the regions of proximity in
resistance units around the focal patches or points.}

\item{\code{summary}}{A summary of the the grains of connectivity generated and
their properties.}

\item{\code{centroids}}{A \code{SpatialPoints} object indicating the grain's polygon
centroids.}

\item{\code{th}}{A list of \code{igraph} objects giving the graphs describing the relationship
                 among the polygons in that grain

See \code{\link{grain}} for more information.}
}}

\author{
Alex Chubaty and Paul Galpern
}
