% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain.R
\docType{methods}
\name{grain}
\alias{grain}
\alias{grain,goc-method}
\title{Extract a grain of connectivity (GOC) tessellation at a given scale}
\usage{
grain(x, ...)

\S4method{grain}{goc}(x, whichThresh, sp = FALSE, ...)
}
\arguments{
\item{x}{A \code{goc} object created by \code{\link{GOC}}.}

\item{...}{Additional arguments to pass to \code{plot} if \code{doPlot = TRUE}).}

\item{whichThresh}{Integer giving the grain threshold to extract.}

\item{sp}{Logical.  If \code{TRUE} then produce a \code{\link{SpatialPolygonsDataFrame}}
representation of the selected threshold.
Requires also running \code{\link{GOC}} with \code{sp = TRUE},
and that the \code{rgeos} package is installed.}
}
\value{
A list object containing the following elements:

\describe{
  \item{\code{summary}}{gives the properties of the visualized scale of the GOC model;}

  \item{\code{voronoi}}{a \code{RasterLayer} giving the Voronoi tessellation;}

  \item{\code{centroids}}{a \code{SpatialPoints} objects giving the centroids
  of the polygons in the tessellation;}

  \item{\code{voronoiSP}}{vector representation of polygons in the tessellation
  (\code{SpatialPolygonsDataFrame}; if \code{sp = TRUE})}
}
}
\description{
Extract a tessellation (i.e., scales) from a GOC model.
}
\examples{
\dontrun{
library(raster)

## Load raster landscape
tiny <- raster(system.file("extdata/tiny.asc", package = "grainscape"))

## Create a resistance surface from a raster using an is-becomes reclassification
tinyCost <- reclassify(tiny, rcl = cbind(c(1, 2, 3, 4), c(1, 5, 10, 12)))

## Produce a patch-based MPG where patches are resistance features=1
tinyPatchMPG <- MPG(cost = tinyCost, patch = (tinyCost == 1))

## Extract a representative subset of 5 grains of connectivity
tinyPatchGOC <- GOC(tinyPatchMPG, nThresh = 5)

## look at the maxLink column for possible whichThresh values
tinyPatchGOC@summary$maxLink ## 0, 30, 50, 64, 114

## Very quick visualization at the finest scale/grain/threshold,
## producing plot on the default graphics device
tinyPatchGOCgrain <- grain(tinyPatchGOC, whichThresh = 0)
plot(tinyPatchGOCgrain, col = topo.colors(10))

## Visualize the model at the finest scale/grain/threshold
## Manual control of plotting
plot(grain(tinyPatchGOC, whichThresh = 0)@voronoi,
     col = sample(rainbow(100)), legend = FALSE, main = "Threshold 1")

## Extract a representative subset of 5 grains of connectivity for vector visualization
tinyPatchGOC <- GOC(tinyPatchMPG, nThresh = 5, sp = TRUE)

## grain the model at a selected scale/grain/threshold using vector polygons
plot(tinyPatchMPG@patchId, col = "grey", legend = FALSE)
plot(grain(tinyPatchGOC, whichThresh = 50, sp = TRUE)@voronoiSP, add = TRUE, lwd = 2)
}

}
\references{
Fall, A., M.-J. Fortin, M. Manseau, D. O'Brien. (2007) Spatial graphs: Principles and applications for habitat connectivity. Ecosystems 10:448:461.

Galpern, P., M. Manseau, P.J. Wilson. (2012) Grains of connectivity: analysis at multiple spatial scales in landscape genetics.  Molecular Ecology 21:3996-4009.
}
\seealso{
\code{\link{GOC}}
}
\author{
Paul Galpern and Alex Chubaty
}
