% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainscape-defunct.R
\name{grainscape-defunct}
\alias{grainscape-defunct}
\alias{gsMPG}
\alias{gsGOC}
\alias{gsGOCCorridor}
\alias{gsGOCDistance}
\alias{gsGOCPoint}
\alias{gsGOCVisualize}
\alias{gsGraphDataFrame}
\alias{gsMPGstitch}
\title{Defunct functions in grainscape}
\usage{
gsMPG(...)

gsGOC(...)

gsGOCCorridor(...)

gsGOCDistance(...)

gsGOCPoint(...)

gsGOCVisualize(...)

gsGraphDataFrame(...)

gsMPGstitch(...)
}
\arguments{
\item{...}{Arguments formerly passed to these functions; now ignored.}
}
\description{
These functions are defunct and have been removed from \pkg{grainscape}.
They were deprecated in version 0.4.0, when the functions were renamed to
drop their \code{gs}/\code{gsGOC} prefixes (#10), and are removed in version 1.0.0.
Use the replacement function named in each function's error message.
}
\keyword{internal}
