% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainscape2-package.R
\docType{data}
\name{grainscape2-maps}
\alias{grainscape2-maps}
\title{Test maps included with \code{grainscape2}}
\format{raster}
\description{
Test maps included with \code{grainscape2}
}
\details{
\code{fragmented.asc}: DESCRIPTION NEEDED
\code{patchy.asc}: DESCRIPTION NEEDED
\code{tiny.asc}: DESCRIPTION NEEDED
}
\keyword{maps}

