% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGOCDistance.R
\docType{methods}
\name{gsGOCDistance}
\alias{gsGOCDistance}
\title{Find the grains of connectivity network distance}
\usage{
gsGOCDistance(gsGOC, coords, weight = "meanWeight")
}
\arguments{
\item{gsGOC}{A \code{gsGOC} object produced by \code{\link{gsGOC}}.}

\item{coords}{A two column matrix or a \code{\link{SpatialPoints}} object
giving the coordinates of points of interest.}

\item{weight}{The GOC graph link weight to use in calculating the distance.
Please see Details for explanation.}
}
\value{
A list object giving a distance matrix for each threshold in the \code{gsGOC} object.
Distance matrices give the pairwise grains of connectivity network distances between sampling locations.
Matrix indices correspond to rows in the \code{coords} matrix.
}
\description{
Given a \code{\link{gsGOC}} object find the shortest network distance between
pairs of points using the GOC graph.  This can be used as an effective distance
for landscape connectivity assessments.
}
\examples{
\dontrun{
## Load raster landscape
tiny <- raster(system.file("extdata/tiny.asc", package = "grainscape2"))

## Create a resistance surface from a raster using an is-becomes reclassifification
tinyCost <- reclassify(tiny, rcl = cbind(c(1, 2, 3, 4), c(1, 5, 10, 12)))

## Produce a patch-based MPG where patches are resistance features=1
tinyPatchMPG <- gsMPG(cost = tinyCost, patch = tinyCost == 1)

## Extract a representative subset of 5 grains of connectivity
tinyPatchGOC <- gsGOC(tinyPatchMPG, nThresh = 5)

## Three sets of coordinates in the study area
loc <- cbind(c(30, 60, 90), c(30, 60, 90))

## Find the GOC network distance matrices between these points
## for each of the 5 grains of connectivity
tinyDist <- gsGOCDistance(tinyPatchGOC, loc)
}

}
\author{
Paul Galpern
}
\references{
Fall, A., M.-J. Fortin, M. Manseau, D. O'Brien.  (2007) Spatial graphs:  Principles and applications for habitat connectivity.  Ecosystems.  10:448:461\cr\cr
Galpern, P., M. Manseau, P.J. Wilson. (2012) Grains of connectivity: analysis at multiple spatial scales in landscape genetics.  Molecular Ecology 21:3996-4009.\cr
}
\seealso{
\code{\link{gsGOC}}, \code{\link{gsGOCPoint}}
}

