% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsHabConnEngine.R
\docType{methods}
\name{habConnEngine}
\alias{habConnEngine}
\title{Habitat connectivity engine}
\usage{
habConnEngine(cost, hab, nodata = as.integer(NA), fpthresh = 1e-04)
}
\arguments{
\item{cost}{Raster cost (resistance) map.}

\item{hab}{Numeric value corresponding to habitat cells.}

\item{nodata}{Numeric value corresponding to 'no data' or \code{NA}.}

\item{fpthresh}{Threshold value for making comparisons of floating point numbers.}
}
\value{
An object of class \code{hce} containing the following components:\cr
        1. \code{voronoi}: a raster whose values indicate the voronoi tesselation;\cr
        2. \code{patchLinks}: a raster whose values indicate patch ids (positive integers)
           and link ids (negative integers);\cr
        3. \code{linkData}: data.frame of link attributes.
}
\description{
DESCRIPTION NEEDED
}
\examples{
cost <- raster(system.file("extdata/fragmented.asc", package = "grainscape2"))
if (interactive()) plot(cost)

# cells in raster `cost` with value of 1 are habitat (patch) cells
# cell in raster `cost` with value -9999 should be interpreted as NA (no data)
links <- habConnEngine(cost, hab = 1, nodata = -9999)

}
\author{
Alex Chubaty
}
\seealso{
\code{link{habConnRcpp}}
}

