% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{.habConnRcpp}
\alias{.habConnRcpp}
\title{Habitat connectivity engine (C++)}
\usage{
.habConnRcpp(cost, nrow, ncol, hab)
}
\arguments{
\item{cost}{A numeric vector of habitat cost (resistance) values
extracted from a raster cost map.}

\item{nrow}{Number of rows in the raster cost/patch map.}

\item{ncol}{Number of columns in the raster cost/patch map.}

\item{hab}{Numeric value corresponding to habitat cells in the cost map.}
}
\description{
Internal function, not intended to be called directly.
DESCRIPTION NEEDED
}
\author{
Sam Doctolero
}
\keyword{internal}

