% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,corridor,ANY-method}
\alias{plot,corridor,ANY-method}
\alias{plot,grain,ANY-method}
\alias{plot,mpg,ANY-method}
\title{Plotting \code{grainscape} objects}
\usage{
\S4method{plot}{corridor,ANY}(x, y, ...)

\S4method{plot}{grain,ANY}(x, y, ...)

\S4method{plot}{mpg,ANY}(x, y, ...)
}
\arguments{
\item{x}{A \code{grainscape} object (\code{corridor}, \code{grain}, or \code{mpg}).}

\item{y}{Ignored.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\description{
Plotting \code{grainscape} objects
}
\seealso{
\code{\linkS4class{corridor}},
         \code{\linkS4class{grain}},
         \code{\linkS4class{mpg}}
}
\author{
Alex Chubaty and Paul Galpern
}
