% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,goc-method}
\alias{show,goc-method}
\title{Show a \code{goc} object}
\usage{
\S4method{show}{goc}(object)
}
\arguments{
\item{object}{A \code{\link[=goc-class]{goc}} object created using \code{\link{GOC}}.}
}
\description{
Custom \code{show} method to safely print the contents of a \code{goc} object.
}
\note{
When a \code{goc} object is created using \code{GOC(..., sp= FALSE)} (the default),
an 'empty' \code{SpatialPointsDataFrame} is stored in the \code{goc}'s \code{voronoiSP}
slot. This object cannot be safely printed, so the default \code{show} method
is bypassed, printing each of the other slots and showing \code{voronoiSP} as \code{<empty>}.
}
