% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{corridor-class}
\alias{corridor-class}
\title{The \code{corridor} class}
\description{
The \code{corridor} class
}
\section{Slots}{

\describe{
\item{\code{voronoi}}{A \code{RasterLayer} representation of the boundaries
of the voronoi polygons.}

\item{\code{linksSP}}{A \code{SpatialLinesDataFrame} representation of links
in the grains of connectivity graph.}

\item{\code{nodesSP}}{A \code{SpatialPoints} representation of the nodes in
the grains of connectivity graph}

\item{\code{shortestLinksSP}}{A \code{SpatialLines} representation of the links in
the shortest path between coordinates}

\item{\code{shortestNodesSP}}{A \code{SpatialPoints} representation of the nodes in
the shortest path between coordinates}

\item{\code{corridorLength}}{A \code{numeric} of length 1 giving the length of the
                       shortest path between coordinates in accumulated
                       resistance units.

See \code{\link{corridor}} for more information.}
}}

\author{
Alex Chubaty and Paul Galpern
}
