% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gsMPGstitch}
\alias{gsMPGstitch}
\title{\code{grainscape}: Defunct}
\usage{
gsMPGstitch(...)
}
\arguments{
\item{...}{Any arguments passed to defunct functions.}
}
\description{
These functions have removed from \code{grainscape}.
}
