% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainscape2-package.R
\docType{data}
\name{grainscape-maps}
\alias{grainscape-maps}
\title{Test maps included with \code{grainscape}}
\format{raster}
\description{
Test maps included with \code{grainscape}
}
\details{
\code{fragmented.asc}: DESCRIPTION NEEDED - PLEASE WRITE
\code{patchy.asc}: DESCRIPTION NEEDED - PLEASE WRITE
\code{tiny.asc}: DESCRIPTION NEEDED - PLEASE WRITE
}
\keyword{maps}

