% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsHabConnEngine.R
\docType{methods}
\name{habConnEngine}
\alias{habConnEngine}
\title{Habitat connectivity engine}
\usage{
habConnEngine(cost, hab, nodata = as.integer(NA), fpthresh = 1e-04)
}
\arguments{
\item{cost}{Raster cost (resistance) map.}

\item{hab}{Numeric value corresponding to habitat cells.}

\item{nodata}{Numeric value corresponding to 'no data' or \code{NA}.}

\item{fpthresh}{Threshold value for making comparisons of floating point numbers.}
}
\description{
DESCRIPTION NEEDED
}
\examples{
cost <- raster(system.file("extdata/fragmented.asc", package = "grainscape2"))
if (interactive()) plot(cost)

# cells in raster `cost` with value of 1 are habitat (patch) cells
# cell in raster `cost` with value -9999 should be interpreted as NA (no data)
habConnEngine(cost, hab = 1, nodata = -9999)

}
\author{
Alex Chubaty
}
\seealso{
\code{link{habConnRcpp}}
}

