% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitatConnectivityEngine.R
\name{.habConnEngine}
\alias{.habConnEngine}
\title{Habitat connectivity engine}
\usage{
.habConnEngine(cost, patches)
}
\arguments{
\item{cost}{Numeric raster cost (resistance) map.}

\item{patches}{Logical raster indicating presence of habitat patches.}
}
\value{
An object of class \code{\linkS4class{hce}}.
}
\description{
Internal function. Serves as a wrapper around the habitat connectivity engine
developed in C++.
}
\examples{
library(raster)

cost <- raster(system.file("extdata/fragmented.asc", package = "grainscape"))
plot(cost)

# cells in raster `cost` with value of 1 are habitat (patch) cells
links <- grainscape:::.habConnEngine(cost, patches = (cost == 1))

links                  # examine the object
plot(links@voronoi)    # plot the voronoi tesselation
plot(links@patchLinks) # plot the patches and links

}
\seealso{
\code{link{habConnRcpp}}
}
\author{
Alex Chubaty
}
\keyword{internal}
