% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsHabConnEngine.R
\docType{methods}
\name{habConnEngine}
\alias{habConnEngine}
\title{Habitat connectivity engine}
\usage{
habConnEngine(cost, patches)
}
\arguments{
\item{cost}{Numeric raster cost (resistance) map.}

\item{patches}{Logical raster indicating presence of habitat patches.}
}
\value{
An object of class \code{hce} containing the following components:\cr
        1. \code{voronoi}: a raster whose values indicate the voronoi tesselation;\cr
        2. \code{patchLinks}: a raster whose values indicate patch ids (positive integers)
           and link ids (negative integers);\cr
        3. \code{linkData}: data.frame of link attributes.
}
\description{
DESCRIPTION NEEDED
}
\examples{
cost <- raster::raster(system.file("extdata/fragmented.asc", package = "grainscape2"))
if (interactive()) plot(cost)

# cells in raster `cost` with value of 1 are habitat (patch) cells
links <- habConnEngine(cost, patches = (cost == 1))


if (interactive()) {
  links                  # examine the object
  plot(links$voronoi)    # plot the voronoi tesselation
  plot(links$patchLinks) # plot the patches and links
}
}
\author{
Alex Chubaty
}
\seealso{
\code{link{habConnRcpp}}
}

