% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{habConnRcpp}
\alias{habConnRcpp}
\title{Habitat connectivity}
\usage{
habConnRcpp(cost, nrow, ncol, hab, no_data, distinctValues, threshold = 1e-04)
}
\arguments{
\item{cost}{DESCRIPTION NEEDED}

\item{nrow}{DESCRIPTION NEEDED}

\item{ncol}{DESCRIPTION NEEDED}

\item{hab}{DESCRIPTION NEEDED}

\item{no_data}{DESCRIPTION NEEDED}

\item{distinctValues}{DESCRIPTION NEEDED}

\item{threshold}{DESCRIPTION NEEDED (default \code{1e-4}).}
}
\description{
DESCRIPTION NEEDED
}
\author{
Sam Doctolero
}

