% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,mpg,ANY-method}
\alias{plot,mpg,ANY-method}
\alias{plot,grain,ANY-method}
\title{Plotting \code{grainscape} objects}
\usage{
\S4method{plot}{mpg,ANY}(x, y, ...)

\S4method{plot}{grain,ANY}(x, y, ...)
}
\arguments{
\item{x}{A \code{mpg} object produced by \code{\link{MPG}} or \code{grain}
object produced by \code{\link{grain}}.}

\item{y}{Ignored.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\description{
Plotting \code{grainscape} objects
}
\examples{
\dontrun{
library(igraph)
library(raster)

## Load raster landscape
tiny <- raster(system.file("extdata/tiny.asc", package = "grainscape"))

## Create a resistance surface from a raster using an is-becomes reclassifyification
tinyCost <- reclassify(tiny, rcl = cbind(c(1, 2, 3, 4), c(1, 5, 10, 12)))

## Produce a patch-based MPG where patches are resistance features=1
tinyPatchMPG <- MPG(cost = tinyCost, patch = (tinyCost == 1))

## Quick visualization of the MPG
plot(tinyPatchMPG, col = c("grey", "black"), legend = FALSE)
}

}
