% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,goc-method}
\alias{show,goc-method}
\alias{show,grain-method}
\alias{show,corridor-method}
\title{Show a \code{grainscape} object}
\usage{
\S4method{show}{goc}(object)

\S4method{show}{grain}(object)

\S4method{show}{corridor}(object)
}
\arguments{
\item{object}{A \code{\link[=goc-class]{goc}} or
\code{\link[=grain-class]{grain}} object.}
}
\description{
Custom \code{show} method to safely print the contents of a \code{goc} or
\code{grain} object.
}
